+++ OLD/DeviceFunctions.cpp
