+++ OLD/DeviceFunctions.h
