+++ OLD/EncoderFunctions.cpp
