+++ OLD/EncoderFunctions.h
